+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0-alpha.22] - 2022-01-17
+
 ### Added
 - Added ``ble.lpw3.AdvertisementData`` class.
 - Added ``ble.lpw3.BootloaderAdvertisementData`` class.
@@ -196,7 +198,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.21..HEAD
+[Unreleased]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.22..HEAD
+[1.0.0-alpha.22]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.21...v1.0.0-alpha.22
 [1.0.0-alpha.21]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.20...v1.0.0-alpha.21
 [1.0.0-alpha.20]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.19...v1.0.0-alpha.20
 [1.0.0-alpha.19]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.18...v1.0.0-alpha.19